Further refactors, KcqlSettings is now minimal
