+++ kafka-connect-cassandra/src/test/scala/io/lenses/streamreactor/common/offsets/TestOffsetHandler.scala
@@ -15,11 +15,15 @@
  */
 package io.lenses.streamreactor.common.offsets
 
-import io.lenses.streamreactor.common.TestUtilsBase
+import org.apache.kafka.connect.source.SourceTaskContext
+import org.apache.kafka.connect.storage.OffsetStorageReader
 import org.mockito.MockitoSugar
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 
+import java.util
+import java.util.Collections
+import scala.jdk.CollectionConverters.MapHasAsJava
 import scala.jdk.CollectionConverters.SeqHasAsJava
 
 /**
@@ -30,7 +34,7 @@ import scala.jdk.CollectionConverters.SeqHasAsJava
   * Created by andrew@datamountaineer.com on 27/04/16.
   * stream-reactor
   */
-class TestOffsetHandler extends AnyWordSpec with Matchers with MockitoSugar with TestUtilsBase {
+class TestOffsetHandler extends AnyWordSpec with Matchers with MockitoSugar {
   "should return an offset" in {
     val lookupPartitionKey = "test_lk_key"
     val offsetValue        = "2013-01-01 00:05+0000"
@@ -44,4 +48,36 @@ class TestOffsetHandler extends AnyWordSpec with Matchers with MockitoSugar with
     val offsetRecovered  = OffsetHandler.recoverOffset[String](offsetsRecovered, lookupPartitionKey, table, offsetColumn)
     offsetRecovered.get shouldBe offsetValue
   }
+
+  private def getSourceTaskContext(
+    lookupPartitionKey: String,
+    offsetValue:        String,
+    offsetColumn:       String,
+    table:              String,
+  ): SourceTaskContext = {
+
+    /**
+      * offset holds a map of map[string, something],map[identifier, value]
+      *
+      * map(map(assign.import.table->table1) -> map("my_timeuuid"->"2013-01-01 00:05+0000")
+      */
+
+    //set up partition
+    val partition: util.Map[String, String] = Collections.singletonMap(lookupPartitionKey, table)
+    //as a list to search for
+    val partitionList: util.List[util.Map[String, String]] = List(partition).asJava
+    //set up the offset
+    val offset: util.Map[String, Object] = Collections.singletonMap(offsetColumn, offsetValue)
+    //create offsets to initialize from
+    val offsets: util.Map[util.Map[String, String], util.Map[String, Object]] = Map(partition -> offset).asJava
+
+    //mock out reader and task context
+    val taskContext = mock[SourceTaskContext]
+    val reader      = mock[OffsetStorageReader]
+    when(reader.offsets(partitionList)).thenReturn(offsets)
+    when(taskContext.offsetStorageReader()).thenReturn(reader)
+
+    taskContext
+  }
+
 }
