+++ kafka-connect-mongodb/src/main/scala/io/lenses/streamreactor/connect/mongodb/config/MongoSettings.scala
@@ -59,7 +59,7 @@ object MongoSettings extends StrictLogging {
     val errorPolicy      = config.getErrorPolicy
     val retries          = config.getNumberRetries
     val rowKeyBuilderMap = config.getUpsertKeys(kcql = kcql, preserveFullKeys = true)
-    val fieldsMap        = getFieldsMap(kcql)
+    val fieldsMap        = config.getFieldsMap(kcql)
     val ignoreFields     = config.getIgnoreFieldsMap()
 
     val username = config.getUsername
@@ -120,7 +120,4 @@ object MongoSettings extends StrictLogging {
     logger.info(s"MongoConfigConstants.JSON_DATETIME_FIELDS_CONFIG is $set")
     set
   }
-
-  def getFieldsMap(kcql: Set[Kcql]): Map[String, Map[String, String]] =
-    kcql.toList.map(rm => (rm.getSource, rm.getFields.asScala.map(fa => (fa.toString, fa.getAlias)).toMap)).toMap
 }
